Rename ingredient order rule
